pyproject.toml
[project]
name = "agentic-ai-workshop"
version = "0.1.0"
description = "agentic-ai-workshop"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb==0.4.24",
    "faiss-cpu==1.8.0",
    "fastapi==0.110.0",
    "fastmcp>=2.14.7",
    "langchain==1.2.15",
    "langchain-community==0.4.1",
    "langchain-google-genai==4.2.1",
    "langchain-ollama>=1.1.0",
    "langchain-openai==1.1.12",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters==1.1.1",
    "openai==2.31.0",
    "pydantic>=2.11.0",
    "pypdf==4.0.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.30.0",
    "wikipedia>=1.4.0",
]
